requirements.txt
flask
requests
PyJWT
Werkzeug